requirements.txt
fastapi==0.98.0
uvicorn[standard]==0.22.0
httpx==0.24.0
asyncpg==0.27.0
pydantic==2.5.0
prometheus-client==0.17.0
python-dotenv==1.0.0
numpy==1.26.5
scikit-learn==1.3.3
tenacity==8.2.2
structlog==24.1.0
slowapi==0.1.7
pgvector==0.9.0
cachetools==5.4.0
faiss-cpu==1.7.4
hdbscan==0.8.35
